# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Medical Appointment System API
  version: 1.0.0
  description: API Documentation for the CareOnline Medical Appointment System
servers:
  - url: http://localhost:5001/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /login:
    post:
      tags: [Auth]
      # ยิงไปเพื่อเช็ค Username/Password แล้วขอรับ Token
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
  /register:
    post:
      tags: [Auth]
      # ยิงข้อมูลไปพื่อสมัครเป็นสมาชิกระดับ User
      summary: Register a new user
      security: [] 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User registered successfully
...